src/test/javascript/spec/app/entities/inventory-item-variance/inventory-item-variance.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { InventoryItemVarianceService } from 'app/entities/inventory-item-variance/inventory-item-variance.service';
import { IInventoryItemVariance, InventoryItemVariance } from 'app/shared/model/inventory-item-variance.model';

describe('Service Tests', () => {
  describe('InventoryItemVariance Service', () => {
    let injector: TestBed;
    let service: InventoryItemVarianceService;
    let httpMock: HttpTestingController;
    let elemDefault: IInventoryItemVariance;
    let expectedResult: IInventoryItemVariance | IInventoryItemVariance[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(InventoryItemVarianceService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new InventoryItemVariance(0, 'AAAAAAA', 0, 0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a InventoryItemVariance', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new InventoryItemVariance()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a InventoryItemVariance', () => {
        const returnedFromService = Object.assign(
          {
            varianceReason: 'BBBBBB',
            atpVar: 1,
            qohVar: 1,
            comments: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of InventoryItemVariance', () => {
        const returnedFromService = Object.assign(
          {
            varianceReason: 'BBBBBB',
            atpVar: 1,
            qohVar: 1,
            comments: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a InventoryItemVariance', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
